pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hpc-map"
version = "0.1.0"
description = "HPC motorway distance map prototype"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "PyYAML>=6.0.1",
]

[tool.setuptools.packages.find]
include = ["backend*", "pipeline*"]
